refactor: Split simple.go into focused, well-named modules
Refactored the 449-line simple.go file into smaller, focused modules:
- chat.go (346 lines): Main chat model and UI rendering
- commands.go (98 lines): Command handling logic
- suggestions.go (35 lines): Autocomplete functionality
- styles.go (14 lines): Color scheme and styling

Benefits:
- Better separation of concerns
- More meaningful file names
- Easier to maintain and extend
- Clearer code organization

Maintained backward compatibility with SimpleModel alias.
